+++ modules/__tests__/Message-test.js
@@ -91,30 +91,6 @@ describe('Message', function () {
     });
   });
 
-  describe('setCookie', function () {
-    describe('when no cookies have been previously set', function () {
-      it('sets the "Set-Cookie" header to the appropriate string', function () {
-        message.setCookie('cookieName', { value: 'cookieValue' });
-        expect(message.headers['Set-Cookie']).toEqual('cookieName=cookieValue');
-      });
-    });
-
-    describe('when cookies have been previously set', function () {
-      beforeEach(function () {
-        message.setCookie('previousOne', { value: 'previousOneValue' });
-      });
-
-      it('sets the "Set-Cookie" header to an array of headers', function () {
-        message.setCookie('cookieName', { value: 'cookieValue' });
-
-        expect(message.headers['Set-Cookie']).toEqual([
-          'previousOne=previousOneValue',
-          'cookieName=cookieValue',
-        ]);
-      });
-    });
-  });
-
   describe('bufferContent', function () {
     it('responds to the message content being set', function () {
       message.content = 'foo';